\name{DailyValues}
\alias{DailyValues}
\docType{data}
\title{
Sample Rain and Wind Data
}
\description{
Sample data from Nov 15 to Jan 13 for obtaining rain values
}
\usage{data("DailyValues")}
\format{
  A data frame with 60 observations on the following 4 variables.
  \describe{
    \item{\code{Day}}{a string vector}
    \item{\code{Month}}{a string vector}
    \item{\code{Rain}}{a numeric vector}
    \item{\code{Wind}}{a string vector}
  }
}
\value{
  dataframe 'sample_data'
}
\arguments{
  \item{max}{Max number of data to return}
}
